Introducing Angle and Point
Point.java represents an coordinate in 2d to 3d space. You can use it as a vector as well. It is mostly a classic location data type, but it has everything you would need. Angle.java represents an angle in degrees. It has useful operations for comparing angles. 20 degrees is less than 150, but greater than 340 for instance. It is however dependent on the Point library since it can create an angle of a point relative to another among other things.